% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrange_solution.R
\name{arrange_solution}
\alias{arrange_solution}
\title{arrange_solution
arrange_solution() gives the solution for all
items of 5.3.1 exercise of R for Data Science
employing the arrange method of dplyr}
\description{
arrange_solution
arrange_solution() gives the solution for all
items of 5.3.1 exercise of R for Data Science
employing the arrange method of dplyr
}
