% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_solution.R
\name{FilterMethod}
\alias{FilterMethod}
\title{filter_solution
A solution for exercise 5.2.4 of R for Data Science
employing the filter method from dplyr.}
\description{
filter_solution
A solution for exercise 5.2.4 of R for Data Science
employing the filter method from dplyr.
}
