% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flight_evaluation.R
\name{FlightEval}
\alias{FlightEval}
\alias{flight_evaluation}
\title{flight_evaluation
A solution for exercise 5.6.7 of R for Data Science}
\usage{
flight_evaluation()
}
\description{
flight_evaluation
A solution for exercise 5.6.7 of R for Data Science
}
\examples{
flight_evaluation()
prints 5 different ways to evaluate the Typical Delay Characteristics of a group of flights.
}
