% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grouped_mutates_solution.R
\name{grouped_mutates_solution}
\alias{grouped_mutates_solution}
\title{grouped_mutates_solution
a solution for exercise 5.7.1 of r for Data Science}
\description{
grouped_mutates_solution
a solution for exercise 5.7.1 of r for Data Science
}
\examples{
grouped_mutates_solution()
prints the solution in data frames
For this item of the exercise, two cases can be presented:
to be able to find the worst plane with respect to the flights,
that were not delayed or cancelled. Part (1)
the second is regarding the delay in the arrival of the plane.part (2)
}
