% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_answer.R
\name{Main}
\alias{Main}
\alias{retrieve_answer}
\title{retrieve_answer
retrieve_answer() receive a parameter to return a solutions from chapter 5
of R for Data Science book and a initial script for prime numbers}
\usage{
retrieve_answer(option)
}
\arguments{
\item{option}{A integer number}
}
\description{
retrieve_answer
retrieve_answer() receive a parameter to return a solutions from chapter 5
of R for Data Science book and a initial script for prime numbers
}
\examples{
retrieve_answer(1)
retrieve_answer(2)
retrieve_answer(3)
retrieve_answer(4)
retrieve_answer(5)
retrieve_answer(6)
retrieve_answer(7)

}
