% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_solution.R
\name{mutate_solution}
\alias{mutate_solution}
\title{mutate_solution
mutate_solution() gives the solution for 5.5.2 items 1 and 2
this method use mutate() method from dlpyr}
\description{
converts time into minutes considering a 24h mins range
}
