% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prime_numbers.R
\name{prime_numbers}
\alias{prime_numbers}
\title{Prime Numbers list.
Gives the list of prime numbers between 1 to 100}
\usage{
prime_numbers()
}
\value{
A list of prime numbers between
}
\description{
Prime Numbers list.
Gives the list of prime numbers between 1 to 100
}
