% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_solution.R
\name{SelectMethod}
\alias{SelectMethod}
\title{select_solution
select_solution() gives the solution for 5.4.1 Items 2, 3 and 4
employing the select method from dplyr.}
\description{
select_solution
select_solution() gives the solution for 5.4.1 Items 2, 3 and 4
employing the select method from dplyr.
}
